--- generators/schema_protocols.py
@@ -1,799 +1,0 @@
-#
-# Generated by "infrahubctl protocols"
-#
-
-from __future__ import annotations
-
-from typing import TYPE_CHECKING, Optional
-
-from infrahub_sdk.protocols import CoreNode, BuiltinIPAddress, BuiltinIPAddressSync, BuiltinIPNamespace, BuiltinIPNamespaceSync, BuiltinIPPrefix, BuiltinIPPrefixSync, BuiltinTag, BuiltinTagSync, CoreAccount, CoreAccountSync, CoreArtifact, CoreArtifactCheck, CoreArtifactCheckSync, CoreArtifactDefinition, CoreArtifactDefinitionSync, CoreArtifactSync, CoreArtifactTarget, CoreArtifactTargetSync, CoreArtifactThread, CoreArtifactThreadSync, CoreArtifactValidator, CoreArtifactValidatorSync, CoreChangeComment, CoreChangeCommentSync, CoreChangeThread, CoreChangeThreadSync, CoreCheck, CoreCheckDefinition, CoreCheckDefinitionSync, CoreCheckSync, CoreComment, CoreCommentSync, CoreCredential, CoreCredentialSync, CoreCustomWebhook, CoreCustomWebhookSync, CoreDataCheck, CoreDataCheckSync, CoreDataValidator, CoreDataValidatorSync, CoreFileCheck, CoreFileCheckSync, CoreFileThread, CoreFileThreadSync, CoreGeneratorCheck, CoreGeneratorCheckSync, CoreGeneratorDefinition, CoreGeneratorDefinitionSync, CoreGeneratorGroup, CoreGeneratorGroupSync, CoreGeneratorInstance, CoreGeneratorInstanceSync, CoreGeneratorValidator, CoreGeneratorValidatorSync, CoreGenericAccount, CoreGenericAccountSync, CoreGenericRepository, CoreGenericRepositorySync, CoreGraphQLQuery, CoreGraphQLQueryGroup, CoreGraphQLQueryGroupSync, CoreGraphQLQuerySync, CoreGroup, CoreGroupSync, CoreIPAddressPool, CoreIPAddressPoolSync, CoreIPPrefixPool, CoreIPPrefixPoolSync, CoreNodeSync, CoreNumberPool, CoreNumberPoolSync, CoreObjectThread, CoreObjectThreadSync, CorePasswordCredential, CorePasswordCredentialSync, CoreProfile, CoreProfileSync, CoreProposedChange, CoreProposedChangeSync, CoreReadOnlyRepository, CoreReadOnlyRepositorySync, CoreRepository, CoreRepositorySync, CoreRepositoryValidator, CoreRepositoryValidatorSync, CoreResourcePool, CoreResourcePoolSync, CoreSchemaCheck, CoreSchemaCheckSync, CoreSchemaValidator, CoreSchemaValidatorSync, CoreStandardCheck, CoreStandardCheckSync, CoreStandardGroup, CoreStandardGroupSync, CoreStandardWebhook, CoreStandardWebhookSync, CoreTaskTarget, CoreTaskTargetSync, CoreThread, CoreThreadComment, CoreThreadCommentSync, CoreThreadSync, CoreTransformJinja2, CoreTransformJinja2Sync, CoreTransformPython, CoreTransformPythonSync, CoreTransformation, CoreTransformationSync, CoreUserValidator, CoreUserValidatorSync, CoreValidator, CoreValidatorSync, CoreWebhook, CoreWebhookSync, InternalAccountToken, InternalAccountTokenSync, InternalRefreshToken, InternalRefreshTokenSync, IpamNamespace, IpamNamespaceSync, LineageOwner, LineageOwnerSync, LineageSource, LineageSourceSync
-
-if TYPE_CHECKING:
-    from infrahub_sdk.node import RelatedNode, RelationshipManager
-
-
-class InfraEndpoint(CoreNode):
-    connected_endpoint: RelatedNode
-    profiles: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-class LocationGeneric(CoreNode):
-    name: str
-    shortname: str
-    description: Optional[str]
-    timezone: Optional[str]
-    prefixes: RelationshipManager
-    vlans: RelationshipManager
-    devices: RelationshipManager
-    network_management_servers: RelationshipManager
-    tags: RelationshipManager
-    policy: RelatedNode
-    profiles: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-    parent: RelatedNode
-    children: RelationshipManager
-class OrganizationGeneric(CoreNode):
-    name: str
-    description: Optional[str]
-    tags: RelationshipManager
-    asn: RelationshipManager
-    profiles: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-class SecurityGenericAddress(CoreNode):
-    name: str
-    address_groups: RelationshipManager
-    profiles: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-class SecurityGenericAddressGroup(CoreNode):
-    name: str
-    description: Optional[str]
-    addresses: RelationshipManager
-    profiles: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-    parent: RelatedNode
-    children: RelationshipManager
-class InfraGenericDevice(CoreNode):
-    name: str
-    description: Optional[str]
-    status: Optional[str]
-    location: RelatedNode
-    interfaces: RelationshipManager
-    asn: RelatedNode
-    tags: RelationshipManager
-    primary_address: RelatedNode
-    device_type: RelatedNode
-    platform: RelatedNode
-    topology: RelatedNode
-    profiles: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-class TopologyGenericElement(CoreNode):
-    name: str
-    description: Optional[str]
-    quantity: int
-    topology: RelatedNode
-    profiles: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-class SecurityGenericService(CoreNode):
-    name: str
-    description: Optional[str]
-    service_groups: RelationshipManager
-    profiles: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-class SecurityGenericServiceGroup(CoreNode):
-    name: str
-    description: Optional[str]
-    services: RelationshipManager
-    profiles: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-    parent: RelatedNode
-    children: RelationshipManager
-class InfraInterface(CoreNode):
-    name: str
-    description: Optional[str]
-    speed: int
-    mtu: Optional[int]
-    enabled: Optional[bool]
-    status: Optional[str]
-    role: Optional[str]
-    device: RelatedNode
-    tags: RelationshipManager
-    profiles: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-class NetworkManagementServer(CoreNode):
-    name: str
-    description: Optional[str]
-    status: str
-    location: RelationshipManager
-    ip_addresses: RelationshipManager
-    profiles: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-class TopologyNetworkService(CoreNode):
-    name: Optional[str]
-    description: Optional[str]
-    status: str
-    identifier: RelatedNode
-    topology: RelatedNode
-    profiles: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-class TopologyNetworkStrategy(CoreNode):
-    name: str
-    description: Optional[str]
-    topology: RelationshipManager
-    profiles: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-class SecurityPolicyAssignment(CoreNode):
-    rules: RelationshipManager
-    profiles: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-
-
-class SecurityAddressGroup(SecurityGenericAddressGroup):
-    name: str
-    description: Optional[str]
-    addresses: RelationshipManager
-    profiles: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-    parent: RelatedNode
-    children: RelationshipManager
-class InfraAutonomousSystem(CoreNode):
-    name: str
-    asn: int
-    description: Optional[str]
-    organization: RelatedNode
-    location: RelatedNode
-    devices: RelationshipManager
-    profiles: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-class InfraBGPPeerGroup(CoreNode):
-    name: str
-    description: Optional[str]
-    import_policies: Optional[str]
-    export_policies: Optional[str]
-    maximum_routes: Optional[int]
-    send_community: Optional[bool]
-    local_as: RelatedNode
-    remote_as: RelatedNode
-    profiles: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-class InfraBGPSession(CoreArtifactTarget):
-    type: str
-    description: Optional[str]
-    import_policies: Optional[str]
-    export_policies: Optional[str]
-    status: str
-    role: str
-    local_as: RelatedNode
-    remote_as: RelatedNode
-    local_ip: RelatedNode
-    remote_ip: RelatedNode
-    device: RelatedNode
-    peer_group: RelatedNode
-    peer_session: RelatedNode
-    artifacts: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-    profiles: RelationshipManager
-class LocationBuilding(LocationGeneric):
-    facility_id: Optional[str]
-    physical_address: Optional[str]
-    name: str
-    shortname: str
-    description: Optional[str]
-    timezone: Optional[str]
-    owner: RelatedNode
-    prefixes: RelationshipManager
-    vlans: RelationshipManager
-    devices: RelationshipManager
-    network_management_servers: RelationshipManager
-    tags: RelationshipManager
-    policy: RelatedNode
-    profiles: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-    parent: RelatedNode
-    children: RelationshipManager
-class InfraCircuit(CoreNode):
-    circuit_id: str
-    description: Optional[str]
-    vendor_id: Optional[str]
-    status: str
-    role: str
-    provider: RelatedNode
-    circuit_type: RelatedNode
-    endpoints: RelationshipManager
-    profiles: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-class InfraCircuitEndpoint(InfraEndpoint):
-    description: Optional[str]
-    location: RelatedNode
-    circuit: RelatedNode
-    connected_endpoint: RelatedNode
-    profiles: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-class InfraCircuitType(CoreNode):
-    name: str
-    description: Optional[str]
-    circuits: RelationshipManager
-    profiles: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-class LocationContinent(LocationGeneric):
-    name: str
-    shortname: str
-    description: Optional[str]
-    timezone: Optional[str]
-    prefixes: RelationshipManager
-    vlans: RelationshipManager
-    devices: RelationshipManager
-    network_management_servers: RelationshipManager
-    tags: RelationshipManager
-    policy: RelatedNode
-    profiles: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-    children: RelationshipManager
-    parent: RelatedNode
-class LocationCountry(LocationGeneric):
-    name: str
-    shortname: str
-    description: Optional[str]
-    timezone: Optional[str]
-    prefixes: RelationshipManager
-    vlans: RelationshipManager
-    devices: RelationshipManager
-    network_management_servers: RelationshipManager
-    tags: RelationshipManager
-    policy: RelatedNode
-    profiles: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-    parent: RelatedNode
-    children: RelationshipManager
-class InfraDevice(CoreArtifactTarget, InfraGenericDevice):
-    role: Optional[str]
-    name: str
-    description: Optional[str]
-    status: Optional[str]
-    artifacts: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-    location: RelatedNode
-    interfaces: RelationshipManager
-    asn: RelatedNode
-    tags: RelationshipManager
-    primary_address: RelatedNode
-    device_type: RelatedNode
-    platform: RelatedNode
-    topology: RelatedNode
-    profiles: RelationshipManager
-class InfraDeviceType(CoreNode):
-    name: str
-    description: Optional[str]
-    part_number: Optional[str]
-    height: Optional[int]
-    full_depth: Optional[bool]
-    weight: Optional[int]
-    platform: RelatedNode
-    manufacturer: RelatedNode
-    tags: RelationshipManager
-    profiles: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-class NetworkDhcpOption(CoreNode):
-    name: str
-    description: Optional[str]
-    option_code: int
-    content: str
-    profiles: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-class NetworkDhcpServer(NetworkManagementServer):
-    lease_time: str
-    name: str
-    description: Optional[str]
-    status: str
-    dhcp_options: RelationshipManager
-    location: RelationshipManager
-    ip_addresses: RelationshipManager
-    profiles: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-class TopologyEVPNStrategy(TopologyNetworkStrategy):
-    underlay: str
-    overlay: str
-    name: str
-    description: Optional[str]
-    topology: RelationshipManager
-    profiles: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-class SecurityFQDN(SecurityGenericAddress):
-    fqdn: str
-    name: str
-    address_groups: RelationshipManager
-    profiles: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-class SecurityFirewall(InfraGenericDevice, CoreArtifactTarget, SecurityPolicyAssignment):
-    role: Optional[str]
-    name: str
-    description: Optional[str]
-    status: Optional[str]
-    policy: RelatedNode
-    location: RelatedNode
-    interfaces: RelationshipManager
-    asn: RelatedNode
-    tags: RelationshipManager
-    primary_address: RelatedNode
-    device_type: RelatedNode
-    platform: RelatedNode
-    topology: RelatedNode
-    artifacts: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-    rules: RelationshipManager
-    profiles: RelationshipManager
-class SecurityFirewallInterface(InfraInterface, InfraEndpoint):
-    name: str
-    description: Optional[str]
-    speed: int
-    mtu: Optional[int]
-    enabled: Optional[bool]
-    status: Optional[str]
-    role: Optional[str]
-    ip_addresses: RelationshipManager
-    security_zone: RelatedNode
-    device: RelatedNode
-    tags: RelationshipManager
-    connected_endpoint: RelatedNode
-    profiles: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-class LocationFloor(LocationGeneric):
-    name: str
-    shortname: str
-    description: Optional[str]
-    timezone: Optional[str]
-    prefixes: RelationshipManager
-    vlans: RelationshipManager
-    devices: RelationshipManager
-    network_management_servers: RelationshipManager
-    tags: RelationshipManager
-    policy: RelatedNode
-    profiles: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-    parent: RelatedNode
-    children: RelationshipManager
-class SecurityIPAMIPAddress(SecurityGenericAddress):
-    description: Optional[str]
-    name: str
-    ip_address: RelatedNode
-    address_groups: RelationshipManager
-    profiles: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-class SecurityIPAMIPPrefix(SecurityGenericAddress):
-    description: Optional[str]
-    name: str
-    ip_prefix: RelatedNode
-    address_groups: RelationshipManager
-    profiles: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-class InfraIPAddress(BuiltinIPAddress):
-    address: str
-    description: Optional[str]
-    interface: RelatedNode
-    ip_namespace: RelatedNode
-    ip_prefix: RelatedNode
-    profiles: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-class SecurityIPAddress(SecurityGenericAddress):
-    address: str
-    description: Optional[str]
-    name: str
-    address_groups: RelationshipManager
-    profiles: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-class SecurityIPProtocol(SecurityGenericService):
-    protocol: Optional[int]
-    name: str
-    description: Optional[str]
-    service_groups: RelationshipManager
-    profiles: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-class SecurityIPRange(SecurityGenericAddress):
-    start: str
-    end: str
-    name: str
-    address_groups: RelationshipManager
-    profiles: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-class InfraInterfaceL2(InfraInterface, InfraEndpoint, CoreArtifactTarget):
-    l2_mode: str
-    name: str
-    description: Optional[str]
-    speed: int
-    mtu: Optional[int]
-    enabled: Optional[bool]
-    status: Optional[str]
-    role: Optional[str]
-    untagged_vlan: RelatedNode
-    tagged_vlan: RelationshipManager
-    device: RelatedNode
-    tags: RelationshipManager
-    connected_endpoint: RelatedNode
-    artifacts: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-    profiles: RelationshipManager
-class InfraInterfaceL3(InfraInterface, InfraEndpoint, CoreArtifactTarget):
-    name: str
-    description: Optional[str]
-    speed: int
-    mtu: Optional[int]
-    enabled: Optional[bool]
-    status: Optional[str]
-    role: Optional[str]
-    ip_addresses: RelationshipManager
-    device: RelatedNode
-    tags: RelationshipManager
-    connected_endpoint: RelatedNode
-    artifacts: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-    profiles: RelationshipManager
-class TopologyLayer2NetworkService(TopologyNetworkService):
-    name: Optional[str]
-    description: Optional[str]
-    status: str
-    vlan: RelatedNode
-    identifier: RelatedNode
-    topology: RelatedNode
-    profiles: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-class TopologyLayer3NetworkService(TopologyNetworkService):
-    name: Optional[str]
-    description: Optional[str]
-    status: str
-    prefix: RelatedNode
-    vlan: RelatedNode
-    identifier: RelatedNode
-    topology: RelatedNode
-    profiles: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-class TopologyMPLSStrategy(TopologyNetworkStrategy):
-    underlay: str
-    overlay: str
-    name: str
-    description: Optional[str]
-    topology: RelationshipManager
-    profiles: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-class OrganizationManufacturer(OrganizationGeneric):
-    name: str
-    description: Optional[str]
-    device_type: RelationshipManager
-    platform: RelationshipManager
-    tags: RelationshipManager
-    asn: RelationshipManager
-    profiles: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-class LocationMetro(LocationGeneric):
-    name: str
-    shortname: str
-    description: Optional[str]
-    timezone: Optional[str]
-    prefixes: RelationshipManager
-    vlans: RelationshipManager
-    devices: RelationshipManager
-    network_management_servers: RelationshipManager
-    tags: RelationshipManager
-    policy: RelatedNode
-    profiles: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-    parent: RelatedNode
-    children: RelationshipManager
-class NetworkNTPServer(NetworkManagementServer):
-    name: str
-    description: Optional[str]
-    status: str
-    location: RelationshipManager
-    ip_addresses: RelationshipManager
-    profiles: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-class NetworkNameServer(NetworkManagementServer):
-    name: str
-    description: Optional[str]
-    status: str
-    location: RelationshipManager
-    ip_addresses: RelationshipManager
-    profiles: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-class TopologyNetworkServiceIdentifier(CoreNode):
-    identifier: int
-    service: RelatedNode
-    profiles: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-class TopologyPhysicalElement(TopologyGenericElement):
-    mtu: Optional[int]
-    border: Optional[bool]
-    mlag_support: Optional[bool]
-    device_role: str
-    name: str
-    description: Optional[str]
-    quantity: int
-    device_type: RelatedNode
-    topology: RelatedNode
-    profiles: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-class InfraPlatform(CoreNode):
-    name: str
-    description: Optional[str]
-    nornir_platform: Optional[str]
-    napalm_driver: Optional[str]
-    netmiko_device_type: Optional[str]
-    ansible_network_os: Optional[str]
-    containerlab_os: Optional[str]
-    devices: RelationshipManager
-    manufacturer: RelatedNode
-    profiles: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-class SecurityPolicy(CoreNode):
-    name: str
-    description: Optional[str]
-    rules: RelationshipManager
-    location_target: RelatedNode
-    device_target: RelatedNode
-    profiles: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-class SecurityPolicyRule(CoreNode):
-    index: int
-    name: str
-    action: Optional[str]
-    log: Optional[bool]
-    policy: RelatedNode
-    source_zone: RelatedNode
-    destination_zone: RelatedNode
-    source_address: RelationshipManager
-    source_groups: RelationshipManager
-    source_services: RelationshipManager
-    source_service_groups: RelationshipManager
-    destination_address: RelationshipManager
-    destination_groups: RelationshipManager
-    destination_services: RelationshipManager
-    destination_service_groups: RelationshipManager
-    profiles: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-class InfraPrefix(BuiltinIPPrefix):
-    status: str
-    role: Optional[str]
-    prefix: str
-    description: Optional[str]
-    member_type: Optional[str]
-    is_pool: Optional[bool]
-    is_top_level: Optional[bool]
-    utilization: Optional[int]
-    netmask: Optional[str]
-    hostmask: Optional[str]
-    network_address: Optional[str]
-    broadcast_address: Optional[str]
-    organization: RelatedNode
-    location: RelatedNode
-    gateway: RelatedNode
-    network_service: RelatedNode
-    vrf: RelatedNode
-    ip_namespace: RelatedNode
-    ip_addresses: RelationshipManager
-    resource_pool: RelationshipManager
-    profiles: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-    parent: RelatedNode
-    children: RelationshipManager
-class SecurityPrefix(SecurityGenericAddress):
-    prefix: str
-    description: Optional[str]
-    name: str
-    address_groups: RelationshipManager
-    profiles: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-class OrganizationProvider(OrganizationGeneric):
-    name: str
-    description: Optional[str]
-    location: RelationshipManager
-    circuit: RelationshipManager
-    tags: RelationshipManager
-    asn: RelationshipManager
-    profiles: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-class LocationRack(LocationGeneric):
-    facility_id: Optional[str]
-    name: str
-    shortname: str
-    description: Optional[str]
-    timezone: Optional[str]
-    owner: RelatedNode
-    prefixes: RelationshipManager
-    vlans: RelationshipManager
-    devices: RelationshipManager
-    network_management_servers: RelationshipManager
-    tags: RelationshipManager
-    policy: RelatedNode
-    profiles: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-    parent: RelatedNode
-    children: RelationshipManager
-class LocationRegion(LocationGeneric):
-    name: str
-    shortname: str
-    description: Optional[str]
-    timezone: Optional[str]
-    prefixes: RelationshipManager
-    vlans: RelationshipManager
-    devices: RelationshipManager
-    network_management_servers: RelationshipManager
-    tags: RelationshipManager
-    policy: RelatedNode
-    profiles: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-    parent: RelatedNode
-    children: RelationshipManager
-class SecurityRenderedPolicyRule(CoreNode):
-    index: int
-    name: str
-    action: Optional[str]
-    log: Optional[bool]
-    source_policy: RelatedNode
-    source_zone: RelatedNode
-    destination_zone: RelatedNode
-    source_address: RelationshipManager
-    source_groups: RelationshipManager
-    source_services: RelationshipManager
-    source_service_groups: RelationshipManager
-    destination_address: RelationshipManager
-    destination_groups: RelationshipManager
-    destination_services: RelationshipManager
-    destination_service_groups: RelationshipManager
-    profiles: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-class InfraRouteTarget(CoreNode):
-    name: str
-    description: Optional[str]
-    vrf: RelationshipManager
-    profiles: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-class SecurityService(SecurityGenericService):
-    port: int
-    name: str
-    description: Optional[str]
-    ip_protocol: RelatedNode
-    service_groups: RelationshipManager
-    profiles: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-class SecurityServiceGroup(SecurityGenericServiceGroup):
-    name: str
-    description: Optional[str]
-    services: RelationshipManager
-    profiles: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-    parent: RelatedNode
-    children: RelationshipManager
-class SecurityServiceRange(SecurityGenericService):
-    start: int
-    end: int
-    name: str
-    description: Optional[str]
-    ip_protocol: RelatedNode
-    service_groups: RelationshipManager
-    profiles: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-class LocationSuite(LocationGeneric):
-    facility_id: Optional[str]
-    name: str
-    shortname: str
-    description: Optional[str]
-    timezone: Optional[str]
-    owner: RelatedNode
-    prefixes: RelationshipManager
-    vlans: RelationshipManager
-    devices: RelationshipManager
-    network_management_servers: RelationshipManager
-    tags: RelationshipManager
-    policy: RelatedNode
-    profiles: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-    parent: RelatedNode
-    children: RelationshipManager
-class OrganizationTenant(OrganizationGeneric):
-    name: str
-    description: Optional[str]
-    location: RelationshipManager
-    circuit: RelationshipManager
-    tags: RelationshipManager
-    asn: RelationshipManager
-    profiles: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-class TopologyTopology(CoreArtifactTarget):
-    name: str
-    description: Optional[str]
-    elements: RelationshipManager
-    strategy: RelatedNode
-    location: RelatedNode
-    devices: RelationshipManager
-    network_services: RelationshipManager
-    artifacts: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-    profiles: RelationshipManager
-class InfraVLAN(CoreNode):
-    name: str
-    description: Optional[str]
-    vlan_id: int
-    status: str
-    role: str
-    location: RelatedNode
-    network_service: RelatedNode
-    profiles: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-class InfraVRF(CoreNode):
-    name: str
-    vrf_rd: Optional[str]
-    import_rt: RelatedNode
-    export_rt: RelatedNode
-    profiles: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
-class SecurityZone(CoreNode):
-    name: str
-    interfaces: RelationshipManager
-    profiles: RelationshipManager
-    member_of_groups: RelationshipManager
-    subscriber_of_groups: RelationshipManager
