--- generators/schema_protocols_sync.py
@@ -1,799 +1,0 @@
-#
-# Generated by "infrahubctl protocols"
-#
-
-from __future__ import annotations
-
-from typing import TYPE_CHECKING, Optional
-
-from infrahub_sdk.protocols import CoreNode, BuiltinIPAddress, BuiltinIPAddressSync, BuiltinIPNamespace, BuiltinIPNamespaceSync, BuiltinIPPrefix, BuiltinIPPrefixSync, BuiltinTag, BuiltinTagSync, CoreAccount, CoreAccountSync, CoreArtifact, CoreArtifactCheck, CoreArtifactCheckSync, CoreArtifactDefinition, CoreArtifactDefinitionSync, CoreArtifactSync, CoreArtifactTarget, CoreArtifactTargetSync, CoreArtifactThread, CoreArtifactThreadSync, CoreArtifactValidator, CoreArtifactValidatorSync, CoreChangeComment, CoreChangeCommentSync, CoreChangeThread, CoreChangeThreadSync, CoreCheck, CoreCheckDefinition, CoreCheckDefinitionSync, CoreCheckSync, CoreComment, CoreCommentSync, CoreCredential, CoreCredentialSync, CoreCustomWebhook, CoreCustomWebhookSync, CoreDataCheck, CoreDataCheckSync, CoreDataValidator, CoreDataValidatorSync, CoreFileCheck, CoreFileCheckSync, CoreFileThread, CoreFileThreadSync, CoreGeneratorCheck, CoreGeneratorCheckSync, CoreGeneratorDefinition, CoreGeneratorDefinitionSync, CoreGeneratorGroup, CoreGeneratorGroupSync, CoreGeneratorInstance, CoreGeneratorInstanceSync, CoreGeneratorValidator, CoreGeneratorValidatorSync, CoreGenericAccount, CoreGenericAccountSync, CoreGenericRepository, CoreGenericRepositorySync, CoreGraphQLQuery, CoreGraphQLQueryGroup, CoreGraphQLQueryGroupSync, CoreGraphQLQuerySync, CoreGroup, CoreGroupSync, CoreIPAddressPool, CoreIPAddressPoolSync, CoreIPPrefixPool, CoreIPPrefixPoolSync, CoreNodeSync, CoreNumberPool, CoreNumberPoolSync, CoreObjectThread, CoreObjectThreadSync, CorePasswordCredential, CorePasswordCredentialSync, CoreProfile, CoreProfileSync, CoreProposedChange, CoreProposedChangeSync, CoreReadOnlyRepository, CoreReadOnlyRepositorySync, CoreRepository, CoreRepositorySync, CoreRepositoryValidator, CoreRepositoryValidatorSync, CoreResourcePool, CoreResourcePoolSync, CoreSchemaCheck, CoreSchemaCheckSync, CoreSchemaValidator, CoreSchemaValidatorSync, CoreStandardCheck, CoreStandardCheckSync, CoreStandardGroup, CoreStandardGroupSync, CoreStandardWebhook, CoreStandardWebhookSync, CoreTaskTarget, CoreTaskTargetSync, CoreThread, CoreThreadComment, CoreThreadCommentSync, CoreThreadSync, CoreTransformJinja2, CoreTransformJinja2Sync, CoreTransformPython, CoreTransformPythonSync, CoreTransformation, CoreTransformationSync, CoreUserValidator, CoreUserValidatorSync, CoreValidator, CoreValidatorSync, CoreWebhook, CoreWebhookSync, InternalAccountToken, InternalAccountTokenSync, InternalRefreshToken, InternalRefreshTokenSync, IpamNamespace, IpamNamespaceSync, LineageOwner, LineageOwnerSync, LineageSource, LineageSourceSync
-
-if TYPE_CHECKING:
-    from infrahub_sdk.node import RelatedNodeSync, RelationshipManagerSync
-
-
-class InfraEndpoint(CoreNode):
-    connected_endpoint: RelatedNodeSync
-    profiles: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-class LocationGeneric(CoreNode):
-    name: str
-    shortname: str
-    description: Optional[str]
-    timezone: Optional[str]
-    prefixes: RelationshipManagerSync
-    vlans: RelationshipManagerSync
-    devices: RelationshipManagerSync
-    network_management_servers: RelationshipManagerSync
-    tags: RelationshipManagerSync
-    policy: RelatedNodeSync
-    profiles: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-    parent: RelatedNodeSync
-    children: RelationshipManagerSync
-class OrganizationGeneric(CoreNode):
-    name: str
-    description: Optional[str]
-    tags: RelationshipManagerSync
-    asn: RelationshipManagerSync
-    profiles: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-class SecurityGenericAddress(CoreNode):
-    name: str
-    address_groups: RelationshipManagerSync
-    profiles: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-class SecurityGenericAddressGroup(CoreNode):
-    name: str
-    description: Optional[str]
-    addresses: RelationshipManagerSync
-    profiles: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-    parent: RelatedNodeSync
-    children: RelationshipManagerSync
-class InfraGenericDevice(CoreNode):
-    name: str
-    description: Optional[str]
-    status: Optional[str]
-    location: RelatedNodeSync
-    interfaces: RelationshipManagerSync
-    asn: RelatedNodeSync
-    tags: RelationshipManagerSync
-    primary_address: RelatedNodeSync
-    device_type: RelatedNodeSync
-    platform: RelatedNodeSync
-    topology: RelatedNodeSync
-    profiles: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-class TopologyGenericElement(CoreNode):
-    name: str
-    description: Optional[str]
-    quantity: int
-    topology: RelatedNodeSync
-    profiles: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-class SecurityGenericService(CoreNode):
-    name: str
-    description: Optional[str]
-    service_groups: RelationshipManagerSync
-    profiles: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-class SecurityGenericServiceGroup(CoreNode):
-    name: str
-    description: Optional[str]
-    services: RelationshipManagerSync
-    profiles: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-    parent: RelatedNodeSync
-    children: RelationshipManagerSync
-class InfraInterface(CoreNode):
-    name: str
-    description: Optional[str]
-    speed: int
-    mtu: Optional[int]
-    enabled: Optional[bool]
-    status: Optional[str]
-    role: Optional[str]
-    device: RelatedNodeSync
-    tags: RelationshipManagerSync
-    profiles: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-class NetworkManagementServer(CoreNode):
-    name: str
-    description: Optional[str]
-    status: str
-    location: RelationshipManagerSync
-    ip_addresses: RelationshipManagerSync
-    profiles: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-class TopologyNetworkService(CoreNode):
-    name: Optional[str]
-    description: Optional[str]
-    status: str
-    identifier: RelatedNodeSync
-    topology: RelatedNodeSync
-    profiles: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-class TopologyNetworkStrategy(CoreNode):
-    name: str
-    description: Optional[str]
-    topology: RelationshipManagerSync
-    profiles: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-class SecurityPolicyAssignment(CoreNode):
-    rules: RelationshipManagerSync
-    profiles: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-
-
-class SecurityAddressGroup(SecurityGenericAddressGroup):
-    name: str
-    description: Optional[str]
-    addresses: RelationshipManagerSync
-    profiles: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-    parent: RelatedNodeSync
-    children: RelationshipManagerSync
-class InfraAutonomousSystem(CoreNode):
-    name: str
-    asn: int
-    description: Optional[str]
-    organization: RelatedNodeSync
-    location: RelatedNodeSync
-    devices: RelationshipManagerSync
-    profiles: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-class InfraBGPPeerGroup(CoreNode):
-    name: str
-    description: Optional[str]
-    import_policies: Optional[str]
-    export_policies: Optional[str]
-    maximum_routes: Optional[int]
-    send_community: Optional[bool]
-    local_as: RelatedNodeSync
-    remote_as: RelatedNodeSync
-    profiles: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-class InfraBGPSession(CoreArtifactTarget):
-    type: str
-    description: Optional[str]
-    import_policies: Optional[str]
-    export_policies: Optional[str]
-    status: str
-    role: str
-    local_as: RelatedNodeSync
-    remote_as: RelatedNodeSync
-    local_ip: RelatedNodeSync
-    remote_ip: RelatedNodeSync
-    device: RelatedNodeSync
-    peer_group: RelatedNodeSync
-    peer_session: RelatedNodeSync
-    artifacts: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-    profiles: RelationshipManagerSync
-class LocationBuilding(LocationGeneric):
-    facility_id: Optional[str]
-    physical_address: Optional[str]
-    name: str
-    shortname: str
-    description: Optional[str]
-    timezone: Optional[str]
-    owner: RelatedNodeSync
-    prefixes: RelationshipManagerSync
-    vlans: RelationshipManagerSync
-    devices: RelationshipManagerSync
-    network_management_servers: RelationshipManagerSync
-    tags: RelationshipManagerSync
-    policy: RelatedNodeSync
-    profiles: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-    parent: RelatedNodeSync
-    children: RelationshipManagerSync
-class InfraCircuit(CoreNode):
-    circuit_id: str
-    description: Optional[str]
-    vendor_id: Optional[str]
-    status: str
-    role: str
-    provider: RelatedNodeSync
-    circuit_type: RelatedNodeSync
-    endpoints: RelationshipManagerSync
-    profiles: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-class InfraCircuitEndpoint(InfraEndpoint):
-    description: Optional[str]
-    location: RelatedNodeSync
-    circuit: RelatedNodeSync
-    connected_endpoint: RelatedNodeSync
-    profiles: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-class InfraCircuitType(CoreNode):
-    name: str
-    description: Optional[str]
-    circuits: RelationshipManagerSync
-    profiles: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-class LocationContinent(LocationGeneric):
-    name: str
-    shortname: str
-    description: Optional[str]
-    timezone: Optional[str]
-    prefixes: RelationshipManagerSync
-    vlans: RelationshipManagerSync
-    devices: RelationshipManagerSync
-    network_management_servers: RelationshipManagerSync
-    tags: RelationshipManagerSync
-    policy: RelatedNodeSync
-    profiles: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-    children: RelationshipManagerSync
-    parent: RelatedNodeSync
-class LocationCountry(LocationGeneric):
-    name: str
-    shortname: str
-    description: Optional[str]
-    timezone: Optional[str]
-    prefixes: RelationshipManagerSync
-    vlans: RelationshipManagerSync
-    devices: RelationshipManagerSync
-    network_management_servers: RelationshipManagerSync
-    tags: RelationshipManagerSync
-    policy: RelatedNodeSync
-    profiles: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-    parent: RelatedNodeSync
-    children: RelationshipManagerSync
-class InfraDevice(CoreArtifactTarget, InfraGenericDevice):
-    role: Optional[str]
-    name: str
-    description: Optional[str]
-    status: Optional[str]
-    artifacts: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-    location: RelatedNodeSync
-    interfaces: RelationshipManagerSync
-    asn: RelatedNodeSync
-    tags: RelationshipManagerSync
-    primary_address: RelatedNodeSync
-    device_type: RelatedNodeSync
-    platform: RelatedNodeSync
-    topology: RelatedNodeSync
-    profiles: RelationshipManagerSync
-class InfraDeviceType(CoreNode):
-    name: str
-    description: Optional[str]
-    part_number: Optional[str]
-    height: Optional[int]
-    full_depth: Optional[bool]
-    weight: Optional[int]
-    platform: RelatedNodeSync
-    manufacturer: RelatedNodeSync
-    tags: RelationshipManagerSync
-    profiles: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-class NetworkDhcpOption(CoreNode):
-    name: str
-    description: Optional[str]
-    option_code: int
-    content: str
-    profiles: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-class NetworkDhcpServer(NetworkManagementServer):
-    lease_time: str
-    name: str
-    description: Optional[str]
-    status: str
-    dhcp_options: RelationshipManagerSync
-    location: RelationshipManagerSync
-    ip_addresses: RelationshipManagerSync
-    profiles: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-class TopologyEVPNStrategy(TopologyNetworkStrategy):
-    underlay: str
-    overlay: str
-    name: str
-    description: Optional[str]
-    topology: RelationshipManagerSync
-    profiles: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-class SecurityFQDN(SecurityGenericAddress):
-    fqdn: str
-    name: str
-    address_groups: RelationshipManagerSync
-    profiles: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-class SecurityFirewall(InfraGenericDevice, CoreArtifactTarget, SecurityPolicyAssignment):
-    role: Optional[str]
-    name: str
-    description: Optional[str]
-    status: Optional[str]
-    policy: RelatedNodeSync
-    location: RelatedNodeSync
-    interfaces: RelationshipManagerSync
-    asn: RelatedNodeSync
-    tags: RelationshipManagerSync
-    primary_address: RelatedNodeSync
-    device_type: RelatedNodeSync
-    platform: RelatedNodeSync
-    topology: RelatedNodeSync
-    artifacts: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-    rules: RelationshipManagerSync
-    profiles: RelationshipManagerSync
-class SecurityFirewallInterface(InfraInterface, InfraEndpoint):
-    name: str
-    description: Optional[str]
-    speed: int
-    mtu: Optional[int]
-    enabled: Optional[bool]
-    status: Optional[str]
-    role: Optional[str]
-    ip_addresses: RelationshipManagerSync
-    security_zone: RelatedNodeSync
-    device: RelatedNodeSync
-    tags: RelationshipManagerSync
-    connected_endpoint: RelatedNodeSync
-    profiles: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-class LocationFloor(LocationGeneric):
-    name: str
-    shortname: str
-    description: Optional[str]
-    timezone: Optional[str]
-    prefixes: RelationshipManagerSync
-    vlans: RelationshipManagerSync
-    devices: RelationshipManagerSync
-    network_management_servers: RelationshipManagerSync
-    tags: RelationshipManagerSync
-    policy: RelatedNodeSync
-    profiles: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-    parent: RelatedNodeSync
-    children: RelationshipManagerSync
-class SecurityIPAMIPAddress(SecurityGenericAddress):
-    description: Optional[str]
-    name: str
-    ip_address: RelatedNodeSync
-    address_groups: RelationshipManagerSync
-    profiles: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-class SecurityIPAMIPPrefix(SecurityGenericAddress):
-    description: Optional[str]
-    name: str
-    ip_prefix: RelatedNodeSync
-    address_groups: RelationshipManagerSync
-    profiles: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-class InfraIPAddress(BuiltinIPAddress):
-    address: str
-    description: Optional[str]
-    interface: RelatedNodeSync
-    ip_namespace: RelatedNodeSync
-    ip_prefix: RelatedNodeSync
-    profiles: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-class SecurityIPAddress(SecurityGenericAddress):
-    address: str
-    description: Optional[str]
-    name: str
-    address_groups: RelationshipManagerSync
-    profiles: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-class SecurityIPProtocol(SecurityGenericService):
-    protocol: Optional[int]
-    name: str
-    description: Optional[str]
-    service_groups: RelationshipManagerSync
-    profiles: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-class SecurityIPRange(SecurityGenericAddress):
-    start: str
-    end: str
-    name: str
-    address_groups: RelationshipManagerSync
-    profiles: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-class InfraInterfaceL2(InfraInterface, InfraEndpoint, CoreArtifactTarget):
-    l2_mode: str
-    name: str
-    description: Optional[str]
-    speed: int
-    mtu: Optional[int]
-    enabled: Optional[bool]
-    status: Optional[str]
-    role: Optional[str]
-    untagged_vlan: RelatedNodeSync
-    tagged_vlan: RelationshipManagerSync
-    device: RelatedNodeSync
-    tags: RelationshipManagerSync
-    connected_endpoint: RelatedNodeSync
-    artifacts: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-    profiles: RelationshipManagerSync
-class InfraInterfaceL3(InfraInterface, InfraEndpoint, CoreArtifactTarget):
-    name: str
-    description: Optional[str]
-    speed: int
-    mtu: Optional[int]
-    enabled: Optional[bool]
-    status: Optional[str]
-    role: Optional[str]
-    ip_addresses: RelationshipManagerSync
-    device: RelatedNodeSync
-    tags: RelationshipManagerSync
-    connected_endpoint: RelatedNodeSync
-    artifacts: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-    profiles: RelationshipManagerSync
-class TopologyLayer2NetworkService(TopologyNetworkService):
-    name: Optional[str]
-    description: Optional[str]
-    status: str
-    vlan: RelatedNodeSync
-    identifier: RelatedNodeSync
-    topology: RelatedNodeSync
-    profiles: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-class TopologyLayer3NetworkService(TopologyNetworkService):
-    name: Optional[str]
-    description: Optional[str]
-    status: str
-    prefix: RelatedNodeSync
-    vlan: RelatedNodeSync
-    identifier: RelatedNodeSync
-    topology: RelatedNodeSync
-    profiles: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-class TopologyMPLSStrategy(TopologyNetworkStrategy):
-    underlay: str
-    overlay: str
-    name: str
-    description: Optional[str]
-    topology: RelationshipManagerSync
-    profiles: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-class OrganizationManufacturer(OrganizationGeneric):
-    name: str
-    description: Optional[str]
-    device_type: RelationshipManagerSync
-    platform: RelationshipManagerSync
-    tags: RelationshipManagerSync
-    asn: RelationshipManagerSync
-    profiles: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-class LocationMetro(LocationGeneric):
-    name: str
-    shortname: str
-    description: Optional[str]
-    timezone: Optional[str]
-    prefixes: RelationshipManagerSync
-    vlans: RelationshipManagerSync
-    devices: RelationshipManagerSync
-    network_management_servers: RelationshipManagerSync
-    tags: RelationshipManagerSync
-    policy: RelatedNodeSync
-    profiles: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-    parent: RelatedNodeSync
-    children: RelationshipManagerSync
-class NetworkNTPServer(NetworkManagementServer):
-    name: str
-    description: Optional[str]
-    status: str
-    location: RelationshipManagerSync
-    ip_addresses: RelationshipManagerSync
-    profiles: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-class NetworkNameServer(NetworkManagementServer):
-    name: str
-    description: Optional[str]
-    status: str
-    location: RelationshipManagerSync
-    ip_addresses: RelationshipManagerSync
-    profiles: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-class TopologyNetworkServiceIdentifier(CoreNode):
-    identifier: int
-    service: RelatedNodeSync
-    profiles: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-class TopologyPhysicalElement(TopologyGenericElement):
-    mtu: Optional[int]
-    border: Optional[bool]
-    mlag_support: Optional[bool]
-    device_role: str
-    name: str
-    description: Optional[str]
-    quantity: int
-    device_type: RelatedNodeSync
-    topology: RelatedNodeSync
-    profiles: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-class InfraPlatform(CoreNode):
-    name: str
-    description: Optional[str]
-    nornir_platform: Optional[str]
-    napalm_driver: Optional[str]
-    netmiko_device_type: Optional[str]
-    ansible_network_os: Optional[str]
-    containerlab_os: Optional[str]
-    devices: RelationshipManagerSync
-    manufacturer: RelatedNodeSync
-    profiles: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-class SecurityPolicy(CoreNode):
-    name: str
-    description: Optional[str]
-    rules: RelationshipManagerSync
-    location_target: RelatedNodeSync
-    device_target: RelatedNodeSync
-    profiles: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-class SecurityPolicyRule(CoreNode):
-    index: int
-    name: str
-    action: Optional[str]
-    log: Optional[bool]
-    policy: RelatedNodeSync
-    source_zone: RelatedNodeSync
-    destination_zone: RelatedNodeSync
-    source_address: RelationshipManagerSync
-    source_groups: RelationshipManagerSync
-    source_services: RelationshipManagerSync
-    source_service_groups: RelationshipManagerSync
-    destination_address: RelationshipManagerSync
-    destination_groups: RelationshipManagerSync
-    destination_services: RelationshipManagerSync
-    destination_service_groups: RelationshipManagerSync
-    profiles: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-class InfraPrefix(BuiltinIPPrefix):
-    status: str
-    role: Optional[str]
-    prefix: str
-    description: Optional[str]
-    member_type: Optional[str]
-    is_pool: Optional[bool]
-    is_top_level: Optional[bool]
-    utilization: Optional[int]
-    netmask: Optional[str]
-    hostmask: Optional[str]
-    network_address: Optional[str]
-    broadcast_address: Optional[str]
-    organization: RelatedNodeSync
-    location: RelatedNodeSync
-    gateway: RelatedNodeSync
-    network_service: RelatedNodeSync
-    vrf: RelatedNodeSync
-    ip_namespace: RelatedNodeSync
-    ip_addresses: RelationshipManagerSync
-    resource_pool: RelationshipManagerSync
-    profiles: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-    parent: RelatedNodeSync
-    children: RelationshipManagerSync
-class SecurityPrefix(SecurityGenericAddress):
-    prefix: str
-    description: Optional[str]
-    name: str
-    address_groups: RelationshipManagerSync
-    profiles: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-class OrganizationProvider(OrganizationGeneric):
-    name: str
-    description: Optional[str]
-    location: RelationshipManagerSync
-    circuit: RelationshipManagerSync
-    tags: RelationshipManagerSync
-    asn: RelationshipManagerSync
-    profiles: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-class LocationRack(LocationGeneric):
-    facility_id: Optional[str]
-    name: str
-    shortname: str
-    description: Optional[str]
-    timezone: Optional[str]
-    owner: RelatedNodeSync
-    prefixes: RelationshipManagerSync
-    vlans: RelationshipManagerSync
-    devices: RelationshipManagerSync
-    network_management_servers: RelationshipManagerSync
-    tags: RelationshipManagerSync
-    policy: RelatedNodeSync
-    profiles: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-    parent: RelatedNodeSync
-    children: RelationshipManagerSync
-class LocationRegion(LocationGeneric):
-    name: str
-    shortname: str
-    description: Optional[str]
-    timezone: Optional[str]
-    prefixes: RelationshipManagerSync
-    vlans: RelationshipManagerSync
-    devices: RelationshipManagerSync
-    network_management_servers: RelationshipManagerSync
-    tags: RelationshipManagerSync
-    policy: RelatedNodeSync
-    profiles: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-    parent: RelatedNodeSync
-    children: RelationshipManagerSync
-class SecurityRenderedPolicyRule(CoreNode):
-    index: int
-    name: str
-    action: Optional[str]
-    log: Optional[bool]
-    source_policy: RelatedNodeSync
-    source_zone: RelatedNodeSync
-    destination_zone: RelatedNodeSync
-    source_address: RelationshipManagerSync
-    source_groups: RelationshipManagerSync
-    source_services: RelationshipManagerSync
-    source_service_groups: RelationshipManagerSync
-    destination_address: RelationshipManagerSync
-    destination_groups: RelationshipManagerSync
-    destination_services: RelationshipManagerSync
-    destination_service_groups: RelationshipManagerSync
-    profiles: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-class InfraRouteTarget(CoreNode):
-    name: str
-    description: Optional[str]
-    vrf: RelationshipManagerSync
-    profiles: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-class SecurityService(SecurityGenericService):
-    port: int
-    name: str
-    description: Optional[str]
-    ip_protocol: RelatedNodeSync
-    service_groups: RelationshipManagerSync
-    profiles: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-class SecurityServiceGroup(SecurityGenericServiceGroup):
-    name: str
-    description: Optional[str]
-    services: RelationshipManagerSync
-    profiles: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-    parent: RelatedNodeSync
-    children: RelationshipManagerSync
-class SecurityServiceRange(SecurityGenericService):
-    start: int
-    end: int
-    name: str
-    description: Optional[str]
-    ip_protocol: RelatedNodeSync
-    service_groups: RelationshipManagerSync
-    profiles: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-class LocationSuite(LocationGeneric):
-    facility_id: Optional[str]
-    name: str
-    shortname: str
-    description: Optional[str]
-    timezone: Optional[str]
-    owner: RelatedNodeSync
-    prefixes: RelationshipManagerSync
-    vlans: RelationshipManagerSync
-    devices: RelationshipManagerSync
-    network_management_servers: RelationshipManagerSync
-    tags: RelationshipManagerSync
-    policy: RelatedNodeSync
-    profiles: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-    parent: RelatedNodeSync
-    children: RelationshipManagerSync
-class OrganizationTenant(OrganizationGeneric):
-    name: str
-    description: Optional[str]
-    location: RelationshipManagerSync
-    circuit: RelationshipManagerSync
-    tags: RelationshipManagerSync
-    asn: RelationshipManagerSync
-    profiles: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-class TopologyTopology(CoreArtifactTarget):
-    name: str
-    description: Optional[str]
-    elements: RelationshipManagerSync
-    strategy: RelatedNodeSync
-    location: RelatedNodeSync
-    devices: RelationshipManagerSync
-    network_services: RelationshipManagerSync
-    artifacts: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-    profiles: RelationshipManagerSync
-class InfraVLAN(CoreNode):
-    name: str
-    description: Optional[str]
-    vlan_id: int
-    status: str
-    role: str
-    location: RelatedNodeSync
-    network_service: RelatedNodeSync
-    profiles: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-class InfraVRF(CoreNode):
-    name: str
-    vrf_rd: Optional[str]
-    import_rt: RelatedNodeSync
-    export_rt: RelatedNodeSync
-    profiles: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
-class SecurityZone(CoreNode):
-    name: str
-    interfaces: RelationshipManagerSync
-    profiles: RelationshipManagerSync
-    member_of_groups: RelationshipManagerSync
-    subscriber_of_groups: RelationshipManagerSync
